Added tests for the lint-files functions 🚦
