+++ src/linters/eslint/lint-files.ts
@@ -1,5 +1,6 @@
 import { loadESLint } from 'eslint'
 
+import colourLog from '@Utils/colour-log'
 import { getCacheDirectory } from '@Utils/cache'
 
 import { processResults } from './process-results'
@@ -7,30 +8,35 @@ import { processResults } from './process-results'
 import type { LintFilesOptions, LintReport } from '@Types/lint'
 
 const lintFiles = async ({ cache, eslintUseLegacyConfig, files, fix }: LintFilesOptions): Promise<LintReport> => {
-  // Load ESLint
-  const ESLint = await loadESLint({
-    useFlatConfig: !eslintUseLegacyConfig,
-  })
+  try {
+    // Load ESLint
+    const ESLint = await loadESLint({
+      useFlatConfig: !eslintUseLegacyConfig,
+    })
 
-  const eslint = new ESLint({
-    cache,
-    cacheLocation: cache ? getCacheDirectory('eslint') : undefined,
-    fix,
-  })
+    const eslint = new ESLint({
+      cache,
+      cacheLocation: cache ? getCacheDirectory('eslint') : undefined,
+      fix,
+    })
 
-  // Run ESLint
-  const results = await eslint.lintFiles(files)
+    // Run ESLint
+    const results = await eslint.lintFiles(files)
 
-  // Process results
-  const report = processResults(results)
+    // Process results
+    const report = processResults(results)
 
-  // Fix files if requested
-  if (fix) {
-    await ESLint.outputFixes(results)
+    // Fix files if requested
+    if (fix) {
+      await ESLint.outputFixes(results)
+    }
+
+    // Return report
+    return report
+  } catch (error) {
+    colourLog.error('An error occurred while running eslint', error)
+    process.exit(1)
   }
-
-  // Return report
-  return report
 }
 
 export default lintFiles
