+++ src/linters/stylelint/lint-files.ts
@@ -1,5 +1,6 @@
 import stylelint from 'stylelint'
 
+import colourLog from '@Utils/colour-log'
 import { getCacheDirectory } from '@Utils/cache'
 
 import { processResults } from './process-results'
@@ -7,29 +8,34 @@ import { processResults } from './process-results'
 import type { LintFilesOptions, LintReport } from '@Types/lint'
 
 const lintFiles = async ({ cache, files, fix }: LintFilesOptions): Promise<LintReport> => {
-  // Run Stylelint
-  const { results, ruleMetadata } = await stylelint.lint({
-    allowEmptyInput: true,
-    cache,
-    cacheLocation: cache ? getCacheDirectory('stylelint') : undefined,
-    config: { // TODO: Replace with externally loaded or user-provided Stylelint config
-      rules: {
-        'declaration-block-no-duplicate-properties': true,
+  try {
+    // Run Stylelint
+    const { results, ruleMetadata } = await stylelint.lint({
+      allowEmptyInput: true,
+      cache,
+      cacheLocation: cache ? getCacheDirectory('stylelint') : undefined,
+      config: { // TODO: Replace with externally loaded or user-provided Stylelint config
+        rules: {
+          'declaration-block-no-duplicate-properties': true,
+        },
       },
-    },
-    files,
-    fix,
-    quietDeprecationWarnings: true,
-    reportDescriptionlessDisables: true,
-    reportInvalidScopeDisables: true,
-    reportNeedlessDisables: true,
-  })
+      files,
+      fix,
+      quietDeprecationWarnings: true,
+      reportDescriptionlessDisables: true,
+      reportInvalidScopeDisables: true,
+      reportNeedlessDisables: true,
+    })
 
-  // Process results
-  const report = processResults(results, ruleMetadata)
+    // Process results
+    const report = processResults(results, ruleMetadata)
 
-  // Return report
-  return report
+    // Return report
+    return report
+  } catch (error) {
+    colourLog.error('An error occurred while running stylelint', error)
+    process.exit(1)
+  }
 }
 
 export default lintFiles
